Rename TestCustomValidator to TestCreateUserCustomValidator
- Update import paths in test files and providers
- Create new TestCreateUserCustomValidator with existing validation logic
- Improve naming consistency for test validator classes
